% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add copy.R
\name{add3}
\alias{add3}
\title{Take the sum of two values}
\usage{
add3(x, y, z)
}
\arguments{
\item{x}{a numeric value}

\item{y}{a numeric value}
}
\value{
the sum of x and y
}
\description{
Take the sum of two values
}
\examples{
add(3, 4)
add(4, 3)
}
