% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathFunctions.R
\name{my_mult}
\alias{my_mult}
\title{My Multiplication}
\usage{
my_mult(x, y)
}
\arguments{
\item{x}{A numeric value to multiply}

\item{y}{A numeric value to multiply}
}
\description{
Returns the product of 2 numbers
}
\examples{
my_mult(4, 5)
}
\keyword{multiply}
\keyword{product}
