% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathFunctions.R
\name{my_sum}
\alias{my_sum}
\title{My Sum}
\usage{
my_sum(x, y)
}
\arguments{
\item{x}{A numeric value to sum}

\item{y}{A numeric value to sum}
}
\description{
Returns the sum of 2 numbers
}
\examples{
my_sum(4, 5)
}
\keyword{add}
